tests/wpt/tests/speculation-rules/prerender/session-history-subframe-navigation.https.html

<!DOCTYPE html>
<title>Test history.length during a subframe navigation</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="../resources/utils.js"></script>
<script src="resources/utils.js"></script>
<script src="resources/session-history-test-util.js"></script>

<body>
<script>
setup(() => assertSpeculationRulesIsSupported());

const uid = token();

promise_test(async () => {
  assert_equals(
    await runTestInPrerender("testSubframeNavigationInPrerender", uid),
    "Passed"
  );
}, "Subframe navigation in prerender replaces the session entry");
</script>
